Fix directory system + CORS proxy issue
