+++ src/components/avatar/Avatar.styles.tsx
