+++ src/components/teams/TeamAvatar.styles.tsx
