+++ src/components/teams/TeamAvatar.tsx
